removed reference to i18n components in admin-ui
