+++ js/apps/admin-ui/package.json
@@ -8,7 +8,8 @@
     ".": {
       "import": "./lib/keycloak-admin-ui.js",
       "types": "./lib/keycloak-admin-ui.d.ts"
-    }
+    },
+    "./styles.css": "./lib/keycloak-admin-ui.css"
   },
   "files": [
     "lib"
