+++ js/apps/admin-ui/src/i18n/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_LOCALE = "en";
