+++ js/apps/admin-ui/src/realm-settings/themes/ThemesTab.tsx
@@ -1,8 +1,5 @@
 import type RealmRepresentation from "@keycloak/keycloak-admin-client/lib/defs/realmRepresentation";
-import { useEnvironment } from "@keycloak/keycloak-ui-shared";
-import { Environment } from "../../environment";
 import { Tab, TabTitleText } from "@patternfly/react-core";
-import JSZip from "jszip";
 import { useTranslation } from "react-i18next";
 import {
   RoutableTabs,
@@ -11,154 +8,18 @@ import {
 import { useRealm } from "../../context/realm-context/RealmContext";
 import useIsFeatureEnabled, { Feature } from "../../utils/useIsFeatureEnabled";
 import { ThemesTabType, toThemesTab } from "../routes/ThemesTab";
-import { LogoContext } from "./LogoContext";
-import { ThemeColors } from "./ThemeColors";
+import { QuickTheme } from "./QuickTheme";
 import { ThemeSettingsTab } from "./ThemeSettings";
-import { BackgroundContext } from "./BackgroundContext";
 
 type ThemesTabProps = {
   realm: RealmRepresentation;
   save: (realm: RealmRepresentation) => void;
 };
 
-export type ThemeRealmRepresentation = RealmRepresentation & {
-  themeName?: string;
-  themeDescription?: string;
-  fileName?: string;
-  favicon?: File;
-  logo?: File;
-  logoWidth?: string;
-  logoHeight?: string;
-  bgimage?: File;
-};
-
 export default function ThemesTab({ realm, save }: ThemesTabProps) {
   const { t } = useTranslation();
   const { realm: realmName } = useRealm();
-  const { environment } = useEnvironment<Environment>();
   const isFeatureEnabled = useIsFeatureEnabled();
-
-  const saveTheme = async (realm: ThemeRealmRepresentation) => {
-    const zip = new JSZip();
-
-    const styles = JSON.parse(realm.attributes?.style ?? "{}");
-    const { favicon, logo, bgimage, fileName, themeName, themeDescription } =
-      realm;
-
-    const name = themeName || "quick-theme";
-    const description = themeDescription || t("themeDescriptionDefault");
-    const logoName =
-      "img/logo" + logo?.name?.substring(logo?.name?.lastIndexOf("."));
-    const bgimageName =
-      "img/bgimage" + bgimage?.name?.substring(bgimage?.name?.lastIndexOf("."));
-
-    if (favicon) {
-      zip.file(`theme/${name}/common/resources/img/favicon.ico`, favicon);
-    }
-    if (logo) {
-      zip.file(`theme/${name}/common/resources/${logoName}`, logo);
-    }
-    if (bgimage) {
-      zip.file(`theme/${name}/common/resources/${bgimageName}`, bgimage);
-    }
-
-    zip.file(
-      `theme/${name}/admin/theme.properties`,
-      `
-parent=keycloak.v2
-import=common/${name}
-
-description=${description}
-
-${logo ? "logo=" + logoName : ""}
-${favicon ? "favIcon=/img/favicon.ico" : ""}
-styles=css/theme-styles.css
-`,
-    );
-
-    zip.file(
-      `theme/${name}/account/theme.properties`,
-      `
-parent=keycloak.v3
-import=common/${name}
-
-description=${description}
-
-${logo ? "logo=" + logoName : ""}
-${favicon ? "favIcon=/img/favicon.ico" : ""}
-styles=css/theme-styles.css
-`,
-    );
-
-    zip.file(
-      `theme/${name}/login/theme.properties`,
-      `
-parent=keycloak.v2
-import=common/${name}
-
-description=${description}
-
-styles=css/styles.css css/theme-styles.css
-`,
-    );
-
-    zip.file(
-      "META-INF/keycloak-themes.json",
-      `{
-  "themes": [{
-      "name" : "${name}",
-      "types": [ "login", "account", "admin", "common" ]
-  }]
-}`,
-    );
-
-    zip.file(
-      "theme-settings.json",
-      JSON.stringify({
-        ...styles,
-        logo: logo ? `theme/${name}/common/resources/${logoName}` : "",
-        bgimage: bgimage ? `theme/${name}/common/resources/${bgimageName}` : "",
-        favicon: favicon
-          ? `theme/${name}/common/resources/img/favicon.ico`
-          : "",
-      }),
-    );
-
-    const toCss = (obj?: object) =>
-      Object.entries(obj || {})
-        .map(([key, value]) => `--pf-v5-global--${key}: ${value};`)
-        .join("\n");
-
-    const loginCss = (
-      await fetch(
-        `/resources/${environment.resourceVersion}/login/keycloak.v2/css/styles.css`,
-      )
-    ).text();
-    zip.file(`theme/${name}/common/resources/css/styles.css`, loginCss);
-
-    zip.file(
-      `theme/${name}/common/resources/css/theme-styles.css`,
-      `:root {
-        ${bgimage ? `--keycloak-bg-logo-url: url('../${bgimageName}');` : ""}
-        ${logo ? `--keycloak-logo-url: url('../${logoName}');` : ""}
-        --keycloak-logo-height: ${realm.logoHeight};
-        --keycloak-logo-width: ${realm.logoWidth};
-        ${toCss(styles.light)}
-      }
-      .pf-v5-theme-dark {
-        ${toCss(styles.dark)}
-      }
-      `,
-    );
-    await zip.generateAsync({ type: "blob" }).then((content) => {
-      const url = URL.createObjectURL(content);
-      const a = document.createElement("a");
-      a.href = url;
-      a.download = fileName || "quick-theme.jar";
-      a.click();
-      URL.revokeObjectURL(url);
-    });
-  };
 
   const param = (tab: ThemesTabType) => ({
     realm: realmName,
@@ -195,11 +56,7 @@ styles=css/styles.css css/theme-styles.css
         data-testid="quickTheme-tab"
         {...quickThemeTab}
       >
-        <BackgroundContext>
-          <LogoContext>
-            <ThemeColors realm={realm} save={saveTheme} />
-          </LogoContext>
-        </BackgroundContext>
+        <QuickTheme realm={realm} />
       </Tab>
     </RoutableTabs>
   );
