+++ js/apps/admin-ui/src/realm-settings/themes/UploadJar.tsx
@@ -2,7 +2,7 @@ import { Button } from "@patternfly/react-core";
 import JSZip from "jszip";
 import { ChangeEvent } from "react";
 import { useTranslation } from "react-i18next";
-import { ThemeRealmRepresentation } from "./ThemesTab";
+import { ThemeRealmRepresentation } from "./QuickTheme";
 
 type UploadJarProps = {
   onUpload: (theme: ThemeRealmRepresentation) => void;
