+++ js/apps/admin-ui/src/realm-settings/user-profile/AttributesGroupForm.tsx
@@ -58,7 +58,7 @@ const defaultValues: FormFields = {
 
 export default function AttributesGroupForm() {
   const { adminClient } = useAdminClient();
-  const { t } = useTranslation();
+  const { t, i18n } = useTranslation();
   const { realm: realmName, realmRepresentation: realm } = useRealm();
   const { config, save } = useUserProfile();
   const navigate = useNavigate();
@@ -113,6 +113,7 @@ export default function AttributesGroupForm() {
             realmName,
             translationsData: { translation },
           });
+          await i18n.reloadResources();
         } catch (error) {
           addError(t("errorSavingTranslations"), error);
         }
