+++ js/apps/admin-ui/src/utils/useLocale.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { useRealm } from "../context/realm-context/RealmContext";
-import { DEFAULT_LOCALE } from "../i18n/i18n";
+import { DEFAULT_LOCALE } from "../i18n/constants";
 
 export default function useLocale() {
   const { realmRepresentation: realm } = useRealm();
